Add option to hide the title, and modify cant2022
This CL adds the poster for the cant2022 concert, and adds an option
which lets a booklet hide the automatic header based on the title and
subtitle (which still appears in the print version, however).

The custom header now no longer appears in the print version, since
we'll be using this to show the concert posters.

Change-Id: Ic51f75fdd6455e990c7381891af1c7541e063698